fixed with floats

+++ payCalculator.py
@@ -4,7 +4,7 @@ def calculatePay():
     # This first line is provided for you
     hrs = input("Enter Hours:")
     rate = input("Enter Rate:")
-    pay = int(hrs)*int(rate)
+    pay = float(hrs)*float(rate)
     print(pay)
     # end assignment
